Fix Run, Stop Buttons. Code cleanup. Add BitmexTest.py. Refactor main.py to Helpers.py. Update README.md
