+++ ExchangeInterfaces/BitmexTest.py
@@ -1,0 +1,5 @@
+from ExchangeInterfaces.BitmexExchange import BitmexExchange
+
+
+class BitmexTest(BitmexExchange):
+    ENDPOINT = "https://testnet.bitmex.com/api/v1"
