+++ README.md
@@ -5,9 +5,10 @@ Copy trading tool for cryptocurrencies - written in Python & Flask
 # Intro
 bot used to make a mass buying or selling of identical bots to do basic copy trading. 
 
-## Supported Exchanges
+#### Supported Exchanges
 - Binance Spot
 - Bitmex
+- Bitmex Testnet
 
 # Installation and Launch
 
@@ -19,20 +20,11 @@ bot used to make a mass buying or selling of identical bots to do basic copy tra
     cp ./config_files/config-sample.json ./config_files/config.json
     ```
 2. Configure `config.json`
-    - Then open `./config_files/config.json` in text editor and paste your api keys and secrets to master and slaves 
+    - Open `./config_files/config.json` in text editor and paste your api keys and secrets to master and slaves 
+    - Possible values for a variable `exchange_name` you can find in folder ExchangeInterfaces. 
 
-    - The value `exchange_name` you can find in folder ExchangeInterfaces
-
-3. Run `api.py`
-    - Windows (PowerShell)
-        
-        ```
-        pyhton api.py
-        ```
-    - Linux 
-    
-        will filled later
-    
+3. Run `pyhton api.py`
+     
 4. Open GUI
     - go to http://127.0.0.1:5000/ or http://0.0.0.0:5000/
     - click `Run` button
@@ -51,8 +43,10 @@ bot used to make a mass buying or selling of identical bots to do basic copy tra
 - Built with bootstrap
 
 # Known Bugs
-- Currently Nothing
-
+- Add and delete slaves buttons not working with new config file. So need to fill `config.json` manually
+- database is not related with `config.json`
+- Bitmex working only with `XBTUSD` and `ETHUSD` pairs now
+ 
  Please [open an issue](https://github.com/MohammedRashad/Crypto-Copy-Trader/issues/new) to help us fix any bugs or request features if needed.
